after uploaded a file should show the list of the files
