+++ ng2-components/ng2-activiti-form/index.ts
@@ -29,11 +29,11 @@ import { ContentWidgetComponent } from './src/components/widgets/content/content
 import { MASK_DIRECTIVE, WIDGET_DIRECTIVES } from './src/components/widgets/index';
 import { WidgetComponent } from './src/components/widgets/widget.component';
 import { ActivitiAlfrescoContentService } from './src/services/activiti-alfresco.service';
-import { ActivitiContentService } from './src/services/activiti-content-service';
 import { EcmModelService } from './src/services/ecm-model.service';
 import { FormRenderingService } from './src/services/form-rendering.service';
 import { FormService } from './src/services/form.service';
 import { NodeService } from './src/services/node.service';
+import { ProcessContentService } from './src/services/process-content.service';
 import { WidgetVisibilityService } from './src/services/widget-visibility.service';
 
 export * from './src/components/form.component';
@@ -41,7 +41,6 @@ export * from './src/components/form-list.component';
 export * from './src/components/widgets/content/content.widget';
 export * from './src/components/start-form.component';
 export * from './src/services/form.service';
-export * from './src/services/activiti-content-service';
 export * from './src/components/widgets/index';
 export * from './src/services/ecm-model.service';
 export * from './src/services/node.service';
@@ -57,14 +56,15 @@ export { FormComponent as ActivitiForm } from './src/components/form.component';
 export { ContentWidgetComponent as ActivitiContent } from './src/components/widgets/content/content.widget';
 export { StartFormComponent as ActivitiStartForm } from './src/components/start-form.component';
 export { DynamicTableRow } from './src/components/widgets/dynamic-table/dynamic-table-row.model';
+export { ProcessContentService as ActivitiContentService } from './src/services/process-content.service';
 
 export const ACTIVITI_FORM_DIRECTIVES: any[] = [
+    ContentWidgetComponent,
+    FormFieldComponent,
     FormComponent,
     FormListComponent,
-    ContentWidgetComponent,
     StartFormComponent,
     StartFormCustomButtonDirective,
-    FormFieldComponent,
     ...WIDGET_DIRECTIVES
 ];
 
@@ -75,13 +75,13 @@ export const DEPRECATED_FORM_DIRECTIVES: any[] = [
 ];
 
 export const ACTIVITI_FORM_PROVIDERS: any[] = [
-    FormService,
-    ActivitiContentService,
-    EcmModelService,
-    NodeService,
-    WidgetVisibilityService,
     ActivitiAlfrescoContentService,
-    FormRenderingService
+    EcmModelService,
+    FormRenderingService,
+    FormService,
+    NodeService,
+    ProcessContentService,
+    WidgetVisibilityService
 ];
 
 @NgModule({
